+++ src/components/Rules/Rules.tsx
@@ -6,44 +6,30 @@ import Icon from 'semantic-ui-react/dist/commonjs/elements/Icon';
 import Input from 'semantic-ui-react/dist/commonjs/elements/Input';
 import Modal from 'semantic-ui-react/dist/commonjs/modules/Modal';
 import useStore from 'store';
-import { TRule } from 'constants/types';
-import { RuleEditor, Share } from 'components';
+import { TRuleEntry } from 'constants/types';
+import { DEFAULT_RULESET_NAMES } from 'constants/constant';
+import { AddRule, RuleContent, RuleEditor, Share } from 'components';
 import styles from './Rules.module.scss';
 
 const Rules: React.FC = () => {
   const rules = useStore(useCallback((state) => state.rules, []));
   const reorderRule = useStore(useCallback((state) => state.reorderRule, []));
-  const addRule = useStore(useCallback((state) => state.addRule, []));
   const darkMode = useStore(useCallback((state) => state.darkMode, []));
   const selectedRuleset = useStore(useCallback((state) => state.selectedRuleset, []));
-  const rulesets = useStore(useCallback((state) => state.rulesets, []));
   const changeRuleset = useStore(useCallback((state) => state.changeRuleset, []));
   const addRuleset = useStore(useCallback((state) => state.addRuleset, []));
   const deleteRuleset = useStore(useCallback((state) => state.deleteRuleset, []));
   const deleteRule = useStore(useCallback((state) => state.deleteRule, []));
-  const [ruleText, setRuleText] = useState('');
   const [rulesetName, setRulesetName] = useState('');
-  const [open, setOpen] = useState(false);
   const [addModal, setAddModal] = useState(false);
   const [confirm, setConfirm] = useState(false);
 
-  const handleReorder = (rule: TRule, i: number, up: boolean) => {
+  const handleReorder = (rule: TRuleEntry, i: number, up: boolean) => {
     if (up) {
       reorderRule(i - 1, rule, i);
     } else {
       reorderRule(i + 1, rule, i);
     }
-  };
-
-  const handleAdd = () => {
-    addRule(ruleText);
-    setOpen(false);
-    setRuleText('');
-  };
-
-  const handleClose = () => {
-    setOpen(false);
-    setRuleText('');
   };
 
   const handleRuleset = (e: React.SyntheticEvent<HTMLElement, Event>, data: DropdownProps) => {
@@ -84,41 +70,7 @@ const Rules: React.FC = () => {
               : ''
           }
         />
-        <Modal
-          closeOnDimmerClick
-          onClose={handleClose}
-          open={open}
-          trigger={
-            <Button
-              className={styles.buttonLarge}
-              color="green"
-              data-testid="add-rule"
-              disabled={!selectedRuleset}
-              inverted={darkMode}
-              onClick={() => setOpen(true)}
-              type="button"
-            >
-              ADD RULE
-              <Icon name="plus" />
-            </Button>
-          }
-        >
-          <Modal.Header>Add Rule</Modal.Header>
-          <Modal.Content style={{ display: 'flex', flexFlow: 'column nowrap', gap: '5px' }}>
-            Please enter the rule description
-            <Input
-              data-testid="add-rule-input"
-              onChange={(e, data) => setRuleText(data.value)}
-              value={ruleText}
-            />
-          </Modal.Content>
-          <Modal.Actions>
-            <Button onClick={handleClose}>Cancel</Button>
-            <Button disabled={ruleText?.length === 0} onClick={handleAdd}>
-              Save
-            </Button>
-          </Modal.Actions>
-        </Modal>
+        <AddRule />
         <Dropdown
           aria-label="rules"
           className={styles.ruleSelect}
@@ -126,7 +78,9 @@ const Rules: React.FC = () => {
           inline
           lazyLoad
           onChange={handleRuleset}
-          options={rulesets}
+          options={Object.keys(rules).map((key) => {
+            return { key, text: key, value: key };
+          })}
           placeholder="Choose a ruleset"
           selection
           value={selectedRuleset}
@@ -158,15 +112,21 @@ const Rules: React.FC = () => {
               onChange={(e, data) => setRulesetName(data.value)}
               value={rulesetName}
             />
+            {!!rules[rulesetName] && (
+              <span style={{ color: 'red' }}>Ruleset with this name already exists</span>
+            )}
           </Modal.Content>
           <Modal.Actions>
             <Button onClick={handleCloseRulesetModal}>Cancel</Button>
-            <Button disabled={rulesetName?.length === 0} onClick={handleAddRuleset}>
+            <Button
+              disabled={rulesetName?.length === 0 || !!rules[rulesetName]}
+              onClick={handleAddRuleset}
+            >
               Save
             </Button>
           </Modal.Actions>
         </Modal>
-        {!!selectedRuleset && selectedRuleset !== '1' && selectedRuleset !== '2' && (
+        {!!selectedRuleset && !DEFAULT_RULESET_NAMES.includes(selectedRuleset) && (
           <Button
             aria-label="deleteruleset"
             className={styles.button}
@@ -190,7 +150,7 @@ const Rules: React.FC = () => {
         />
       </div>
       <div className={styles.rules} data-testid="ruleslist">
-        {rules[selectedRuleset].map((rule, i) => {
+        {rules[selectedRuleset]?.map((rule, i) => {
           return (
             <div className={styles.rule} key={`rule-${i + 1}`}>
               <div className={styles.reorder}>
@@ -217,22 +177,25 @@ const Rules: React.FC = () => {
                   </Button>
                 )}
               </div>
-              <span className={styles.number}>{`${i + 1}.`}</span>
-              <p>{rule.content}</p>
+              <RuleContent i={i} rule={rule} />
               <div className={styles.buttons}>
-                <Button
-                  aria-label="delete rule"
-                  basic
-                  compact
-                  data-testid={`delete-rule-${i}`}
-                  icon
-                  inverted={darkMode}
-                  onClick={() => deleteRule(i)}
-                  type="button"
-                >
-                  <Icon name="trash" />
-                </Button>
-                <RuleEditor content={rule.content} index={i} />
+                {!rule.default && (
+                  <>
+                    <Button
+                      aria-label="delete rule"
+                      basic
+                      compact
+                      data-testid={`delete-rule-${i}`}
+                      icon
+                      inverted={darkMode}
+                      onClick={() => deleteRule(i)}
+                      type="button"
+                    >
+                      <Icon name="trash" />
+                    </Button>
+                    <RuleEditor content={rule.content} index={i} type={rule.type} />
+                  </>
+                )}
               </div>
             </div>
           );